requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-filter==2.2.0
django-markdown-deux==1.0.5
django-pagedown==2.0.3
djangorestframework==3.10.3
Markdown==3.1.1
markdown2==2.3.8
Pillow==6.1.0
pytz==2019.2
sqlparse==0.3.0
